FREEPBX-10885 - Implement sane backup limits.
This ensures that the machine will never try to send
an email larger than 25mb no matter what, and additionally
sets the default size to be sane (10mb, was 25).

Additonally, when the email exceeds the size, this also
sends an email ALERTING the user of the error, rather than
just silently ignoring.