FREEPBX-18160 Fix server migration
Note:

I tested each class as called and items added as expected.
It seems I was building the data array inappropriately so the data
passed was incorrect.

QA NOTE:

This can only be reproduced upgrading to backup 15. You should be able
to just load a mysql dump of backup_servers and backup_items